Created Controllers for all models and redid routes
